deleting unnecessary files
