+++ src/RowHeader.js
@@ -8,8 +8,10 @@ function RowHeader(props) {
     return (react_1.default.createElement("div", { className: 'rowHeader' },
         react_1.default.createElement("div", { className: 'headerName-api' }, props.api),
         react_1.default.createElement("div", { className: 'headerName' }, props.location),
-        react_1.default.createElement("div", { className: 'headerName-status' }, props.status),
-        react_1.default.createElement("div", { className: 'headerName-api' }, props.stable),
+        react_1.default.createElement("div", { className: 'headerName-status' },
+            react_1.default.createElement("div", null, props.status),
+            react_1.default.createElement("button", { className: "filter-button" })),
+        react_1.default.createElement("div", { className: 'headerName-stable' }, props.stable),
         react_1.default.createElement("div", { className: 'headerName-notes' }, props.notes)));
 }
 exports.default = RowHeader;
